requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
chardet>=5.2.0
fpdf2>=2.7.7
langchain-google-genai>=0.0.6
langchain-text-splitters>=0.0.1
langchain==0.3.7 
langchain-community==0.3.7 
langchain-core==0.3.19 
langchain-chroma==0.1.4 
pypdf 
chromadb
